% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data (random variable) from a given distribution}
\usage{
generate_data(
  distribution,
  n,
  mean = NULL,
  sd = NULL,
  size = NULL,
  prob = NULL
)
}
\arguments{
\item{distribution}{The probability distribution.}

\item{n}{The number of values to draw from the probability distribution.}

\item{mean}{The mean (central tendency).}

\item{sd}{The standard deviation (variability).}

\item{size}{The target number of successful trials.}

\item{prob}{The probability of success in each independent and identical
trial.}
}
\value{
A vector of random deviates whose length is \code{n}
}
\description{
The parameters of the expected random variable are used to draw a random
sample from the given probability distribution.
}
\examples{
generate_data("normal", 6, 2.5, 1.3)
generate_data("poisson", 6, 2.5)
generate_data("logistic", 6, prob = 0.5, size = 1)
generate_data("negbin", 6, 2.5, size = 1)
}
