% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powrunr-package.R
\docType{package}
\name{powrunr-package}
\alias{powrunr}
\alias{powrunr-package}
\title{powrunr: Power Analysis Through Simulation}
\description{
Generate simulated data and calculate statistical power for planned experiments.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
